requirements.txt
requests
feedparser
newspaper3k
deep-translator
